removed unnecessary base classes of category-bundle entities
