--- src/Sulu/Bundle/CategoryBundle/Entity/BaseCategoryTranslation.php
@@ -1,186 +1,0 @@
-<?php
-
-/*
- * This file is part of Sulu.
- *
- * (c) MASSIVE ART WebServices GmbH
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
-namespace Sulu\Bundle\CategoryBundle\Entity;
-
-use Sulu\Component\Security\Authentication\UserInterface;
-
-/**
- * CategoryTranslation.
- */
-abstract class BaseCategoryTranslation implements CategoryTranslationInterface
-{
-    /**
-     * @var string
-     */
-    protected $translation;
-
-    /**
-     * @var string
-     */
-    protected $locale;
-
-    /**
-     * @var int
-     */
-    protected $id;
-
-    /**
-     * @var CategoryInterface
-     */
-    protected $category;
-
-    /**
-     * @var UserInterface
-     */
-    protected $creator;
-
-    /**
-     * @var UserInterface
-     */
-    protected $changer;
-
-    /**
-     * @var \DateTime
-     */
-    protected $created;
-
-    /**
-     * @var \DateTime
-     */
-    protected $changed;
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setTranslation($translation)
-    {
-        $this->translation = $translation;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getTranslation()
-    {
-        return $this->translation;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setLocale($locale)
-    {
-        $this->locale = $locale;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getLocale()
-    {
-        return $this->locale;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setCategory(CategoryInterface $category)
-    {
-        $this->category = $category;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getCategory()
-    {
-        return $this->category;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getCreator()
-    {
-        return $this->creator;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setCreator($creator)
-    {
-        $this->creator = $creator;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getChanger()
-    {
-        return $this->changer;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setChanger($changer)
-    {
-        $this->changer = $changer;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getCreated()
-    {
-        return $this->created;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setCreated($created)
-    {
-        $this->created = $created;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getChanged()
-    {
-        return $this->changed;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setChanged($changed)
-    {
-        $this->changed = $changed;
-    }
-}
