--- src/Sulu/Bundle/CategoryBundle/Entity/BaseKeyword.php
@@ -1,172 +1,0 @@
-<?php
-
-/*
- * This file is part of Sulu.
- *
- * (c) MASSIVE ART WebServices GmbH
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
-namespace Sulu\Bundle\CategoryBundle\Entity;
-
-use Sulu\Component\Security\Authentication\UserInterface;
-
-/**
- * The keywords can describe a category with different words.
- */
-abstract class BaseKeyword implements KeywordInterface
-{
-    /**
-     * @var int
-     */
-    protected $id;
-
-    /**
-     * @var string
-     */
-    protected $keyword;
-
-    /**
-     * @var string
-     */
-    protected $locale;
-
-    /**
-     * @var UserInterface
-     */
-    protected $creator;
-
-    /**
-     * @var UserInterface
-     */
-    protected $changer;
-
-    /**
-     * @var \DateTime
-     */
-    protected $created;
-
-    /**
-     * @var \DateTime
-     */
-    protected $changed;
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setLocale($locale)
-    {
-        $this->locale = $locale;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getLocale()
-    {
-        return $this->locale;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setKeyword($keyword)
-    {
-        $this->keyword = $keyword;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getKeyword()
-    {
-        return $this->keyword;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getCreator()
-    {
-        return $this->creator;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setCreator($creator)
-    {
-        $this->creator = $creator;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getChanger()
-    {
-        return $this->changer;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setChanger($changer)
-    {
-        $this->changer = $changer;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getCreated()
-    {
-        return $this->created;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setCreated($created)
-    {
-        $this->created = $created;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getChanged()
-    {
-        return $this->changed;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setChanged($changed)
-    {
-        $this->changed = $changed;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function equals(KeywordInterface $keyword)
-    {
-        return $keyword->getKeyword() === $this->getKeyword()
-            && $keyword->getLocale() === $this->getLocale();
-    }
-}
