+++ src/Sulu/Bundle/CategoryBundle/Entity/Keyword.php
@@ -13,12 +13,48 @@ namespace Sulu\Bundle\CategoryBundle\Entity;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use Sulu\Component\Security\Authentication\UserInterface;
 
 /**
  * The keywords can describe a category with different words.
  */
-class Keyword extends BaseKeyword
+class Keyword implements KeywordInterface
 {
+    /**
+     * @var int
+     */
+    protected $id;
+
+    /**
+     * @var string
+     */
+    protected $keyword;
+
+    /**
+     * @var string
+     */
+    protected $locale;
+
+    /**
+     * @var UserInterface
+     */
+    protected $creator;
+
+    /**
+     * @var UserInterface
+     */
+    protected $changer;
+
+    /**
+     * @var \DateTime
+     */
+    protected $created;
+
+    /**
+     * @var \DateTime
+     */
+    protected $changed;
+
     /**
      * @var Collection
      */
@@ -30,6 +66,123 @@ class Keyword extends BaseKeyword
     public function __construct()
     {
         $this->categoryTranslations = new ArrayCollection();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setLocale($locale)
+    {
+        $this->locale = $locale;
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getLocale()
+    {
+        return $this->locale;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setKeyword($keyword)
+    {
+        $this->keyword = $keyword;
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getKeyword()
+    {
+        return $this->keyword;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getCreator()
+    {
+        return $this->creator;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setCreator($creator)
+    {
+        $this->creator = $creator;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getChanger()
+    {
+        return $this->changer;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setChanger($changer)
+    {
+        $this->changer = $changer;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getCreated()
+    {
+        return $this->created;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setCreated($created)
+    {
+        $this->created = $created;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getChanged()
+    {
+        return $this->changed;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setChanged($changed)
+    {
+        $this->changed = $changed;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function equals(KeywordInterface $keyword)
+    {
+        return $keyword->getKeyword() === $this->getKeyword()
+        && $keyword->getLocale() === $this->getLocale();
     }
 
     /**
