--- src/Sulu/Bundle/CategoryBundle/Resources/config/doctrine/BaseCategoryMeta.orm.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-
-    <mapped-superclass name="Sulu\Bundle\CategoryBundle\Entity\BaseCategoryMeta">
-
-        <id name="id" type="integer" column="id">
-            <generator strategy="AUTO"/>
-        </id>
-
-        <field name="key" column="meta_key" type="string" length="45"/>
-        <field name="value" column="value" type="string" length="255"/>
-        <field name="locale" column="locale" type="string" length="5" nullable="true"/>
-
-        <many-to-one field="category" target-entity="Sulu\Bundle\CategoryBundle\Entity\CategoryInterface" inversed-by="meta">
-            <cascade>
-                <cascade-all/>
-            </cascade>
-            <join-columns>
-                <join-column name="idCategories" on-delete="CASCADE" referenced-column-name="id" nullable="false"/>
-            </join-columns>
-        </many-to-one>
-
-    </mapped-superclass>
-</doctrine-mapping>
