--- src/Sulu/Bundle/CategoryBundle/Resources/config/doctrine/BaseKeyword.orm.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-
-    <mapped-superclass name="Sulu\Bundle\CategoryBundle\Entity\BaseKeyword">
-        <indexes>
-            <index columns="keyword"/>
-        </indexes>
-
-        <unique-constraints>
-            <unique-constraint columns="keyword,locale" name="unique_idx"/>
-        </unique-constraints>
-
-        <id name="id" type="integer" column="id">
-            <generator strategy="AUTO"/>
-        </id>
-
-        <field name="locale" column="locale" type="string" length="5" nullable="false"/>
-        <field name="keyword" column="keyword" type="string" length="255" nullable="false"/>
-    </mapped-superclass>
-</doctrine-mapping>
