+++ src/Sulu/Bundle/CategoryBundle/Resources/config/doctrine/Category.orm.xml
@@ -1,10 +1,32 @@
 <?xml version="1.0" encoding="utf-8"?>
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd"
+                  xmlns:gedmo="http://gediminasm.org/schemas/orm/doctrine-extensions-mapping">
 
     <entity name="Sulu\Bundle\CategoryBundle\Entity\Category" table="ca_categories">
+        <id name="id" type="integer" column="id">
+            <generator strategy="AUTO"/>
+        </id>
 
+        <field name="key" column="category_key" type="string" length="255" nullable="true" unique="true"/>
+        <field name="defaultLocale" column="default_locale" type="string" length="5" nullable="false"/>
+        <field name="lft" type="integer" column="lft">
+            <gedmo:tree-left/>
+        </field>
+        <field name="rgt" type="integer" column="rgt">
+            <gedmo:tree-right/>
+        </field>
+        <field name="depth" type="integer" column="depth">
+            <gedmo:tree-level/>
+        </field>
+
+        <many-to-one field="parent" target-entity="Sulu\Bundle\CategoryBundle\Entity\CategoryInterface" inversed-by="children">
+            <join-columns>
+                <join-column name="idCategoriesParent" referenced-column-name="id" on-delete="CASCADE"/>
+            </join-columns>
+            <gedmo:tree-parent/>
+        </many-to-one>
         <one-to-many field="meta" target-entity="Sulu\Bundle\CategoryBundle\Entity\CategoryMetaInterface" mapped-by="category">
             <cascade>
                 <cascade-all/>
@@ -16,7 +38,8 @@
                 <cascade-all/>
             </cascade>
         </one-to-many>
-
         <one-to-many field="children" target-entity="Sulu\Bundle\CategoryBundle\Entity\CategoryInterface" mapped-by="parent"/>
+
+        <gedmo:tree type="nested"/>
     </entity>
 </doctrine-mapping>
