+++ src/Sulu/Bundle/CategoryBundle/Resources/config/doctrine/CategoryMeta.orm.xml
@@ -4,5 +4,21 @@
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
     <entity name="Sulu\Bundle\CategoryBundle\Entity\CategoryMeta" table="ca_category_meta">
+        <id name="id" type="integer" column="id">
+            <generator strategy="AUTO"/>
+        </id>
+
+        <field name="key" column="meta_key" type="string" length="45"/>
+        <field name="value" column="value" type="string" length="255"/>
+        <field name="locale" column="locale" type="string" length="5" nullable="true"/>
+
+        <many-to-one field="category" target-entity="Sulu\Bundle\CategoryBundle\Entity\CategoryInterface" inversed-by="meta">
+            <cascade>
+                <cascade-all/>
+            </cascade>
+            <join-columns>
+                <join-column name="idCategories" on-delete="CASCADE" referenced-column-name="id" nullable="false"/>
+            </join-columns>
+        </many-to-one>
     </entity>
 </doctrine-mapping>
