+++ src/Sulu/Bundle/CategoryBundle/Resources/config/doctrine/CategoryTranslation.orm.xml
@@ -4,6 +4,21 @@
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
     <entity name="Sulu\Bundle\CategoryBundle\Entity\CategoryTranslation" table="ca_category_translations">
+        <id name="id" type="integer" column="id">
+            <generator strategy="AUTO"/>
+        </id>
+
+        <field name="translation" column="translation" type="string" length="255"/>
+        <field name="locale" column="locale" type="string" length="5"/>
+
+        <many-to-one field="category" target-entity="Sulu\Bundle\CategoryBundle\Entity\CategoryInterface" inversed-by="translations">
+            <cascade>
+                <cascade-all/>
+            </cascade>
+            <join-columns>
+                <join-column name="idCategories" on-delete="CASCADE" referenced-column-name="id" nullable="false"/>
+            </join-columns>
+        </many-to-one>
         <many-to-many field="keywords" mapped-by="categoryTranslations"
                       target-entity="Sulu\Bundle\CategoryBundle\Entity\KeywordInterface"/>
     </entity>
