+++ src/Sulu/Bundle/CategoryBundle/Resources/config/doctrine/Keyword.orm.xml
@@ -4,6 +4,21 @@
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
     <entity name="Sulu\Bundle\CategoryBundle\Entity\Keyword" table="ca_keywords">
+        <indexes>
+            <index columns="keyword"/>
+        </indexes>
+
+        <unique-constraints>
+            <unique-constraint columns="keyword,locale" name="unique_idx"/>
+        </unique-constraints>
+
+        <id name="id" type="integer" column="id">
+            <generator strategy="AUTO"/>
+        </id>
+
+        <field name="locale" column="locale" type="string" length="5" nullable="false"/>
+        <field name="keyword" column="keyword" type="string" length="255" nullable="false"/>
+
         <many-to-many target-entity="Sulu\Bundle\CategoryBundle\Entity\CategoryTranslationInterface" field="categoryTranslations"
                       inversed-by="keywords">
             <join-table name="ca_category_translations_keywords">
