+++ src/Sulu/Bundle/CategoryBundle/Resources/config/serializer/Keyword.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <serializer>
-    <class name="Sulu\Bundle\CategoryBundle\Entity\KeywordInterface" exclusion-policy="ALL">
+    <class name="Sulu\Bundle\CategoryBundle\Entity\Keyword" exclusion-policy="ALL">
         <virtual-property method="getId" name="id" expose="true"/>
         <virtual-property method="getKeyword" name="keyword" expose="true"/>
         <virtual-property method="getLocale" name="locale" expose="true"/>
